Add files via upload
In this task, we were asked to transmit a song using Amplitude modulation and also receive it. The carrier wave frequency was 1 Mega Herz and the challenge was to calculate the sampling rate of the carrier wave and the sampling rate of blocks used while giving the user control over modulation index.